Update and rename Schizophrenic.cs to Paranoia.cs
